backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2
transformers==4.35.2
torch==2.1.1
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
python-dotenv==1.0.0
Pillow==10.1.0
cryptography==41.0.7
gunicorn==21.2.0
django-filter==23.5
pytest==7.4.3
pytest-django==4.7.0
faker==20.1.0